requirements.txt
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
distro==1.7.0
Django==4.0.3
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
idna==3.3
inflection==0.5.1
jsonschema==3.2.0
mccabe==0.6.1
paramiko==2.10.3
phonenumbers==8.12.45
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2022.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
uritemplate==4.1.1